remove app things from swing gui.
